Make gen_repository Python 2/Python 3 compatible
Fedora systems are getting to the point of not shipping Python 2 at all;
for now it is still available from the repositories but not shipped by
default. Also, /usr/bin/python isn't guaranteed to be Python 2 any more:
python-unversioned-command on Fedora 31 ships this as Python 3.

Make gen_repository.py Python 3 compatible and set that as the default
in the shebang, but retain Python 2 compatibility.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>