Fix AES for CMAC fuzzing
Prior to this commit:

 - libFuzzer would reach Botan::OpCMAC just fine,
 - libFuzzer wouldn't find NSS:OpCMAC (or, wouldn't ever issue a
   completely valid OpCMAC call).

This occurred even on a custom corpus with vastly reduced operation,
cipher, and digest list.

With these changes, libFuzzer + Cryptofuzz should be able to
differentially fuzz NSS and Botan (since both trigger on the same
CF_CIPHER now) and both quickly begin producing valid CMAC calls.

What I don't understand is _why_ the AES bool flag is so important: the
helper "IsAES" doesn't appear to be called anywhere, and the bool flag
.AES also doesn't appear to be accessed. This leaves me stumped why this
change is required.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>