Add PSA Crypto module (#65)
* Include cstddef for size_t

This is necessary to build natively with Clang on Ubuntu, which uses libstdc++.

* Update Mbed TLS description

Update the spelling, GitHub organization and config script.

* Create a module for TF-PSA-Crypto

TF-PSA-Crypto uses the Mbed TLS implementation of the PSA Crypto API
specification. This currently means Mbed TLS >= 2.22, and will be
TF-PSA-Crypto in the future (once that is broken out of Mbed TLS 4.0).

The module could be generalized to other PSA implementations, but at
this point it uses some Mbed TLS specific functions related to memory
management.

As of this commit, the module does not yet support any cryptographic
operations. This will come in subsequent commits.

* PSA crypto: support digest calculation

* Avoid space in module name

* Remove redundant check

If the hash length is wrong, a built-in test will catch that.

* Use assertions if something fails

Using CF_CHECK meant that if a call failed, the test case would be
ignored. Use CF_ASSERT to indicate that if a call failed, that is a
failure that the fuzzer must report.

* Use assert macro for psa_crypto_init failure

* Register the PSA-Crypto module